move FixedOrderMap
