+++ clients/src/test/java/org/apache/kafka/common/utils/FixedOrderMapTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.kafka.streams.processor.internals;
+package org.apache.kafka.common.utils;
 
 import org.hamcrest.CoreMatchers;
 import org.junit.Test;
